Do not require version pin for hatch extensions
Many recipies have hatch extensions such as hatch-vcs in their build stanza. It doesn't
seem to make sense to exclude hatch from pins but require the extensions to pinned.

I chose to use a regex to exclude all packages with the 'hatch-' prefix. It might be
overbroad, but proabaly is lower maintainance. An alternative implementation would be to
add the known plugins to PYTHON_BUILD_TOOLS, at the cost of a periodic whack-a-mole
expedition as new plugins are added.

Signed-off-by: Mark Anderson <manderson@anaconda.com>