Adds support for `no_git_on_windows` auto-fix
Adds support for auto fixing the `no-git-on-windows` rule using the new parse
tree tooling in `percy`.

NOTE: This requires changes found in [this percy PR](https://github.com/anaconda-distribution/percy/pull/63)
so DO NOT MERGE, until

This PR is being put up as a proof of concept, to show that the new parse
tree work can be integrated into Anaconda Linter.

`percy` will need a new version bump AND the linter will need to pin to that
new version of `percy` prior to accepting this PR.

Tested this manually. I can add a unit test in a subsequent commit. Again, this
is mostly going out prematurely as a POC.