Fully enforces `pylint` on the project
There should be no net logic changes with this PR, with the exception of fixes
to previously existing logic mistakes, found by the linter.

Initially I was not planning to perform all of these at once. But it turns
out that there weren't nearly as many linting issues as I previously imagined.

So before we greatly expand this project with additional engineering resources,
I wanted to enforce some rules now.

Notes:
- Expands class name regex to allow for snake-casing, to be consistent with
  linting rule names
- Linter is fully enabled in `pre-commit` and `make` commands for all `*.py`
  files (with some exceptions around the `docs/` and other non-critical files)
- Fixes a number of logic errors
- Fixes a number of other linting errors